+++ test/client/Query.test.tsx
@@ -690,6 +690,28 @@ describe('Query component', () => {
       });
     });
 
+    it('skip', done => {
+      const Component = () => (
+        <Query query={allPeopleQuery} skip>
+          {result => {
+            catchAsyncError(done, () => {
+              expect(result.loading).toBeFalsy();
+              expect(result.data).toBe(undefined);
+              expect(result.error).toBe(undefined);
+              done();
+            });
+            return null;
+          }}
+        </Query>
+      );
+
+      wrapper = mount(
+        <MockedProvider mocks={allPeopleMocks} addTypename={false}>
+          <Component />
+        </MockedProvider>,
+      );
+    });
+
     it('onCompleted with data', done => {
       const mocks = [
         {